software-map: move resolved-data input signatures to their own module
Agent-Session: 01a067ce-34b7-7952-a023-ced44207d9ca
